+++ src/alphabet/TournerSensHoraire.java
@@ -3,20 +3,20 @@ import java.util.LinkedList;
 
 public class TournerSensHoraire extends Mouvement {
 	private static final Character REPRESENTATION_BASE='-';
-	private static final double VALEUR_BASE=90;
+	private static final int VALEUR_BASE=90;
 	
-	public TournerSensHoraire(Character representation, LinkedList<MembreAlpha> evolution ,double valeur)  {
+	public TournerSensHoraire(Character representation, LinkedList<MembreAlpha> evolution ,int valeur)  {
 		super(representation,evolution,valeur);
 		this.tourne=true;
 	}
-	public TournerSensHoraire(Character representation, double valeur) {
+	public TournerSensHoraire(Character representation, int valeur) {
 		super(representation,valeur);
 		this.tourne=true;
 	}
 	public TournerSensHoraire() {
 		this(REPRESENTATION_BASE,VALEUR_BASE);
 	}
-	public TournerSensHoraire(double valeur) {
+	public TournerSensHoraire(int valeur) {
 		this(REPRESENTATION_BASE,valeur);
 	}
