+++ src/alphabet/TournerSensTrigo.java
@@ -3,20 +3,20 @@ import java.util.LinkedList;
 
 public class TournerSensTrigo extends Mouvement {
 	private static final Character REPRESENTATION_BASE='+';
-	private static final double VALEUR_BASE=90;
+	private static final int VALEUR_BASE=90;
 	
-	public TournerSensTrigo(Character representation, LinkedList<MembreAlpha> evolution ,double valeur)  {
+	public TournerSensTrigo(Character representation, LinkedList<MembreAlpha> evolution ,int valeur)  {
 		super(representation,evolution,valeur);
 		this.tourne=true;
 	}
-	public TournerSensTrigo(Character representation, double valeur) {
+	public TournerSensTrigo(Character representation, int valeur) {
 		super(representation,valeur);
 		this.tourne=true;
 	}
 	public TournerSensTrigo() {
 		this(REPRESENTATION_BASE,VALEUR_BASE);
 	}
-	public TournerSensTrigo(double valeur) {
+	public TournerSensTrigo(int valeur) {
 		this(REPRESENTATION_BASE,valeur);
 	}
